# [xy] Improve performance for saving block outputs