# Summary
<!-- Brief summary of what your code does -->
Improve performance for saving block outputs
* Not save null data
* Consolidate multiple micro message into one message.

This reduced the time for updating the test pipeline from ~2.2s to ~700ms, and reduced the time of loading the test pipeline from ~1.7s to ~600ms

# Tests
<!-- How did you test your change? -->
tested locally

original messages
<img width="1190" alt="image" src="https://user-images.githubusercontent.com/80284865/212874250-a16909d4-f6f4-4ea8-b929-11c98d51676b.png">
consolidated message
<img width="1152" alt="image" src="https://user-images.githubusercontent.com/80284865/212874306-0d08462c-00c6-4e19-8119-525e0b272a09.png">


cc:
<!-- Optionally mention someone to let them know about this pull request -->